Add confirmation dialog for blog deletion in DeleteBlog.js

+++ blogs-frontend/src/components/DeleteBlog.js
@@ -4,18 +4,21 @@ import '../DeleteBlog.css';
 const DeleteBlog = ({ blogId, onDelete }) => {
   const handleDelete = () => {
     const csrftoken = document.cookie.match('(^|;)\\s*csrftoken\\s*=\\s*([^;]+)')?.pop();
-    axios.delete(`/api/blog/${blogId}/delete/`, {
-      headers: {
-        'X-CSRFToken': csrftoken,
-      },
-    })
-    .then(response => {
-      console.log('Blog deleted successfully', response.data)
-      onDelete()
-    })
-    .catch(error => {
-      console.error('Error deleting blog: ', error)
-    })
+
+    if(window.confirm('Are you sure you want to delete this blog?')) {
+      axios.delete(`/api/blog/${blogId}/delete/`, {
+        headers: {
+          'X-CSRFToken': csrftoken,
+        },
+      })
+      .then(response => {
+        console.log('Blog deleted successfully', response.data)
+        onDelete()
+      })
+      .catch(error => {
+        console.error('Error deleting blog: ', error)
+      })
+    }
   }
 
   return (
